requirements.txt
endee
sentence-transformers
flask
python-dotenv
PyPDF2
werkzeug
